# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workspaces": {
      "post": {
        "operationId": "WorkspaceController_create",
        "summary": "Create a new workspace",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "WorkspaceController_findAll",
        "summary": "Get all workspaces for current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/workspaces/{id}": {
      "get": {
        "operationId": "WorkspaceController_findOne",
        "summary": "Get workspace by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "WorkspaceController_update",
        "summary": "Update workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "WorkspaceController_delete",
        "summary": "Delete workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/workspaces/{id}/members": {
      "post": {
        "operationId": "WorkspaceController_addMember",
        "summary": "Add member to workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member added successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User or workspace not found"
          },
          "409": {
            "description": "User is already a member"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "WorkspaceController_getMembers",
        "summary": "Get workspace members",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of members"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/workspaces/{id}/members/{memberId}": {
      "patch": {
        "operationId": "WorkspaceController_updateMemberRole",
        "summary": "Update member role",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member role updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "WorkspaceController_removeMember",
        "summary": "Remove member from workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/admin/workspaces/{id}": {
      "patch": {
        "operationId": "AdminWorkspaceController_update",
        "summary": "Update any workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminWorkspaceController_delete",
        "summary": "Delete any workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkspaceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Elbruso API",
    "description": "BI System REST API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Workspace name",
            "example": "My Workspace",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "example": "Team collaboration workspace",
            "maxLength": 500
          },
          "icon": {
            "type": "string",
            "description": "Workspace icon (emoji or URL)",
            "example": "🚀"
          },
          "is_template": {
            "type": "boolean",
            "description": "Is this workspace a template",
            "example": false
          }
        },
        "required": [
          "name"
        ]
      },
...
      "WorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "is_template": {
            "type": "boolean"
          },
          "owner_id": {
            "type": "string"
          },
          "owner_email": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "organization_id": {
            "type": "number",
            "nullable": true
          },
          "sport_id": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "is_template",
          "owner_id",
          "owner_email",
          "created_at",
          "updated_at",
          "is_active",
          "metadata",
          "organization_id",
          "sport_id"
        ]
      },
      "WorkspacesListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponseDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Workspace name",
            "example": "Updated Workspace",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "maxLength": 500
          },
          "icon": {
            "type": "string",
            "description": "Workspace icon"
          },
          "is_template": {
            "type": "boolean",
            "description": "Is this workspace a template"
          }
        }
      },
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Member email address",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "description": "Member role",
            "enum": [
              "owner",
              "editor",
              "viewer"
            ],
            "example": "editor"
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "New member role",
            "enum": [
              "owner",
              "editor",
              "viewer"
            ],
            "example": "editor"
          }
        },
        "required": [
          "role"
        ]
      },
...
      "DeleteWorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}